backend/requirements.txt
Django==5.0.1
channels==4.0.0
channels-redis==4.1.0
djangorestframework==3.14.0
daphne==4.0.0
python-dotenv==1.0.0
caldav==1.3.9
icalendar==5.0.11
requests==2.31.0
Pillow==10.1.0
django-cors-headers==4.3.1